// src/app/models/rep_stockexpRPT.ts
export class RepStockExpRPT {
    CNTSTK_TOT: number;
    CNTSTK_20: number;
    CNTSTK_40: number;
    CNTSTK_20ST: number;
    CNTSTK_20OT: number;
    CNTSTK_40ST: number;
    CNTSTK_40HC: number;
    CNTSTK_40OT: number;
    CNTSTK_ME15: number;
    CNTSTK_1530: number;
    CNTSTK_MA30: number;
    CNTSTK_ALR: number;
    CNTSTK_ALA: number;
    CNTSTK_ALV: number;
    CNTSTK_ALG: number;
    CNTSTK_EN: any;
  }