// src/app/models/det_repstockexpcli.ts
export class DetRepStockExpCliRQT {   
    IDUSer: number;
    IDRol: number;
    IdCliente: string;
    }

export class DetRepStockExpCliRPT {   
   
    Nave: string;
    NomLinea: string;
    NomCliente: string;
    Volante: string;
    BlMaster: string;
    Bl: string;
    Contenedor: string;
    Size: number;
    FechaIngreso: string;
  
   
    }
    
        